Fix KeyError in fromarray by specifying dtype=np.uint8

+++ initi.py
@@ -6,7 +6,7 @@ import os, datetime
 def init_and_create_templates(font: str):
     btpng = os.path.join(BASE_DIR, "black-template.png")
     new_dir = os.path.join(BASE_DIR, "fonts", font[:-4])  # cut out .ttf
-    sqr = np.zeros((TEMPLATE_SIZE, TEMPLATE_SIZE, 3))
+    sqr = np.zeros((TEMPLATE_SIZE, TEMPLATE_SIZE, 3), dtype=np.uint8)
     i = Image.fromarray(sqr, "RGB")
     i.save(btpng)
     try:
